feat: modularize TUI dashboard components into individual view files and implement network activity graph and download list UI
